More Map Changes
Chanelog:
Added Hatay and Bahrain as nations that can be released.
Edited the map. Oman borders are now more accurate, especially with the United Kingdom.
Crossable straits added so that troops from Qatar and Saudi Arabia can cross to Bahrain.
Yemen state is renamed to Sana'a.
Colors of Yemen and Turkey are adjusted.